refactor: move nx helpers to tooling-common
