+++ packages/@clc/nx/executors/project-refs/executor.ts
@@ -6,12 +6,9 @@ import type {Executor} from '@nx/devkit';
 import {glob} from 'glob';
 import ts from 'typescript';
 
-import {
-  extractProjectRoot,
-  findLocalPackages,
-  readPackageJson,
-  writePrettierFile,
-} from '../..';
+import {writePrettierFile} from '@code-like-a-carpenter/tooling-common';
+
+import {extractProjectRoot, findLocalPackages, readPackageJson} from '../..';
 
 import type {ProjectRefsExecutor} from './schema';
