+++ packages/@clc/nx/package.json
@@ -5,12 +5,12 @@
     "@babel/generator": "^7.23.6",
     "@babel/template": "^7.23.9",
     "@babel/types": "^7.23.9",
+    "@code-like-a-carpenter/tooling-common": "*",
     "@nx/devkit": "^18.0.2",
     "@schemastore/package": "^0.0.10",
     "esbuild": "^0.20.0",
     "find-up": "^5.0.0",
     "glob": "^10.3.10",
-    "json-schema-to-typescript": "^13.1.2",
     "minimatch": "^9.0.3",
     "prettier": "^3.2.5",
     "remark": "^15.0.1",
