+++ packages/@code-like-a-carpenter/tooling-common/src/index.ts
@@ -1,2 +1,4 @@
+export * from './filesystem';
 export * from './find-local-packages';
 export * from './get-stack-name';
+export * from './json-schema';
